requirements.txt
exif==1.6.0
aiofiles==23.1.0
aiohttp==3.8.4
python-dotenv==0.21.1
vk-api==11.9.9